Updated to include new functions
Updated to include help button that displays instructions to user and sees if a result occurred twice in a row. If the user rolled the same number twice in a row, the color of the number changes so the user knows they did indeed roll and it wasn't a glitch.